+++ presets/react-validation-eloquent/database/migrations/20240315074428_create_users_table.ts
@@ -9,7 +9,7 @@ export const up = function (db: Knex) {
   return db.schema.createTable(table_name, (t) => {
     t.bigIncrements("id").primary().notNullable();
     t.string("email").unique().notNullable();
-    t.string("user_name").notNullable();
+    t.string("username").notNullable();
     t.string("password").notNullable();
     t.string("first_name").nullable();
     t.string("last_name").nullable();