+++ presets/react-validation-eloquent/resources/components/Input.tsx
@@ -1,26 +1,34 @@
-import { errors } from "@lunoxjs/view/client";
+import { usePage } from "@lunoxjs/view-plugin-react";
 
+interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  label?: string;
+}
 const Input = ({
   name = "",
   label = "",
   defaultValue = undefined,
+  value = "",
+  onChange,
   type = "text",
   placeholder = "",
-}) => {
+}: InputProps) => {
+  const { props } = usePage();
   return (
     <div className="mb-3 flex flex-col">
       <label htmlFor={name} className="text-sm text-gray-800">
         {label || name}
       </label>
       <input
+        onChange={onChange}
         type={type}
         name={name}
+        value={value}
         defaultValue={defaultValue}
         className="border border-yellow-200 p-2 rounded focus:outline-yellow"
         placeholder={placeholder}
       />
-      {errors(name) && (
-        <small className="text-red">{errors(name).message} </small>
+      {props?.errors?.[name] && (
+        <small className="text-red">{props.errors[name].message} </small>
       )}
     </div>
   );